Fix bug where we try to sort a generator
To reproduce:
> ykman oath code -H

Issue:
We try to run creds.sort() - however, if the `show hidden` flag is on,
then we never change the generator into a list, causing an error.

Solution:
Change `creds` into a list immediately.

Why not other solutions?

I would have preferred `creds = list(controller.calculate_all())`.
However, using list comprehension rather than the list constructor is
used when we list all credentials, hence matching the style of other
code.

(It's worth noting that because we called a function `list` and thus
overrode the list constructor, running `list()` on any generator was
going to fail.)

We could have also used `sorted()` instead of `creds.sort()`. However,
the code for finding the longest name assumes that it's a list not a
generator as otherwise, it will exhaust the generator and leave nothing
to output.