Strip out spaces from OATH base32 code entry
Some services (e.g. Dropbox, Google) will have spaces in the base32
code. They do it because base32 does not contain spaces and it makes it
easier for a human to read and type.

In the case of ykman, it's likely that folks copy and paste the code and
thus will also copy over the spaces. As it is safe to ignore the spaces,
we should.

@dagheyman recommended we do this inside the parse_b32_key function. I
agree this is the right place.