Remove try-except from PivController.get_{pin,puk}_tries()
This was used in order to capture the error status from `send_cmd`, and
failed when `send_cmd` returned an empty bytestring and did not raise an
error (which happened if `get_pin_retries` was called after the PIN had
been verified).

By passing the argument `check=None` instead, we always get the status
as the second return value.