Make set_access_code and delete_access_code detect silent failures
Some YubiKey firmware versions after 4.3.1 but before 4.3.6 incorrectly
report success after performing the set_access_code command, even though
the new access code was not set on the slot.

Therefore, the set_access_code and delete_access_code methods will now
verify that the correct access code is in fact set on the slot. This is
done by attempting to set the new access code again, but without sending
the current access code argument.

If this fails, then there must be an access code set on the slot, and
the method uses the same trick again to verify that that access code is
the expected one.

If this succeeds, then there must not be an access code set on the slot,
which means the operation silently failed.