Fix how the context is resolved
On Python3 dict.items() returns a view object and changing dictionary
values while iterating this view may lead to unexpected results when we
resolve the context like an attempt to resolve the same object twice.

This patch fixes this by creating a copy of the underlying dict in the
resolve() method.