Migrate to pytest
pytest has a ["fixtures" framework][1] for parameterizing tests and
preparing test environments. This allows us to get rid of most of the
custom test multiplication framework, and return to conventional Python
test syntax with test classes as top-level module members.

This does have the drawback that tests that don't apply to a given
YubiKey are now marked as skipped instead of not appearing at all,
meaning we'll always have "false positive" test skips. I believe the
advantages more than make up for this, though, and we might be able to
improve on the skipping mechanism in the future.

[1]: https://docs.pytest.org/en/3.0.1/fixture.html