Generate key for TOTP when generate is chosen.
Currently, for the "otp chalresp" command, the tool allows the user to specify
both generate and TOTP, but it ignores the generate and prompts the user for a
key, anyhow.

This change checks generate by itself and generates the key if it's true. If
TOTP is chosen, it outputs the generated key as Base32, since that's what it
prompts the user for if generate isn't chosen.