Honor digits even for challenge
The ykchalresp tool will output a six or eight digit OTP code even if
challenge mode is chosen. This change does that for the ykman tool.

In order to know if the user has selected six or eight digits, I added
a zero selection to digits (although the text still says six). Later, if
TOTP mode is selected and digits is zero, we change the zero to six.

This way, behavior is the same when TOTP is selected, or no digit
count is specified.