Send SIGHUP to `yubikey-agent` if it's running and the connection fails
`yubikey-agent` is an SSH agent written in Go that uses the PIV module
of a YubiKey for SSH[1].

Since it takes a persistent transaction on the YubiKey, using e.g.
`ykman` will fail when it's active. I think it's tedious to always send
a SIGHUP to it whenever I want to use `ykman`, so I added this small
patch that does that for me. Inspired by what `age-plugin-yubikey` is
also doing[2].

Not sure about Windows support, but I left it out because `yubikey-agent`
doesn't do any Windows AFAIK and SIGHUP isn't supported there.

[1] https://github.com/FiloSottile/yubikey-agent
[2] https://github.com/str4d/age-plugin-yubikey/commit/1913838f8ed4b30c756c3c20ea5fdf1680ea97ca